requirements.txt
abel
asgiref==3.8.1
Babel==2.15.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.3.2
defusedxml==0.6.0
Django==2.2.13
django-allauth
django-crispy-forms==1.9.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-modeltranslation==0.15.1
django-multiselectfield==0.1.13
django-oscar==2.0.4
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
purl==1.6
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2